#599 - Added madge to detect future circular dependencies
