+++ package.json
@@ -124,6 +124,7 @@
     "karma-mocha-reporter": "2.2.5",
     "karma-webkit-launcher": "2.1.0",
     "license-report": "6.3.0",
+    "madge": "^6.1.0",
     "mkdirp": "1.0.4",
     "mocha": "10.1.0",
     "mockdate": "3.0.5",
@@ -153,6 +154,7 @@
     "compile-validators": "node ./build/compile-validators.js",
     "lint": "eslint . --ext .ts --max-warnings 0",
     "lint:fix": "eslint . --ext .ts --fix",
+    "madge": "madge --circular ./src/index.ts",
     "test:node": "npm run compile-validators && tsc && c8 mocha \"dist/esm/tests/**/*.spec.js\"",
     "test:browser": "npm run compile-validators && cross-env karma start karma.conf.cjs",
     "test:browser-debug": "npm run compile-validators && cross-env karma start karma.conf.debug.cjs",
