+++ src/types/records-types.ts
@@ -48,6 +48,11 @@ export type RecordsWriteMessage = GenericMessage & {
   encryption?: EncryptionProperty;
 };
 
+/**
+ * records with a data size below a threshold are stored within MessageStore with their data embedded
+ */
+export type RecordsWriteMessageWithOptionalEncodedData = RecordsWriteMessage & { encodedData?: string };
+
 export type EncryptionProperty = {
   algorithm: EncryptionAlgorithm;
   initializationVector: string;
