+++ tests/handlers/records-write.spec.ts
@@ -1,7 +1,7 @@
 import type { EncryptionInput } from '../../src/interfaces/records-write.js';
 import type { GenerateFromRecordsWriteOut } from '../utils/test-data-generator.js';
 import type { ProtocolDefinition } from '../../src/types/protocols-types.js';
-import type { RecordsWriteMessageWithOptionalEncodedData } from '../../src/store/storage-controller.js';
+import type { RecordsWriteMessageWithOptionalEncodedData } from '../../src/types/records-types.js';
 import type { DataStore, EventLog, MessageStore } from '../../src/index.js';
 
 import anyoneCollaborateProtocolDefinition from '../vectors/protocol-definitions/anyone-collaborate.json' assert { type: 'json' };
