+++ tests/store/message-store.spec.ts
@@ -3,7 +3,6 @@ import type { RecordsWriteMessage } from '../../src/types/records-types.js';
 
 import { expect } from 'chai';
 
-import { constructRecordsWriteIndexes } from '../../src/handlers/records-write.js';
 import { DidKeyResolver } from '../../src/index.js';
 import { lexicographicalCompare } from '../../src/utils/string.js';
 import { Message } from '../../src/core/message.js';
@@ -229,7 +228,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}]);
@@ -249,7 +248,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}], { messageTimestamp: SortOrder.Ascending });
           expect(messageQuery.length).to.equal(messages.length);
@@ -267,7 +266,7 @@ export function testMessageStore(): void {
             dateCreated: TestDataGenerator.randomTimestamp(),
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}], { dateCreated: SortOrder.Ascending });
@@ -287,7 +286,7 @@ export function testMessageStore(): void {
             dateCreated: TestDataGenerator.randomTimestamp(),
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}], { dateCreated: SortOrder.Descending });
@@ -308,7 +307,7 @@ export function testMessageStore(): void {
             datePublished : TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}], { datePublished: SortOrder.Ascending });
@@ -329,7 +328,7 @@ export function testMessageStore(): void {
             datePublished : TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: messageQuery } = await messageStore.query(alice.did, [{}], { datePublished: SortOrder.Descending });
@@ -351,7 +350,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const { messages: limitQuery } = await messageStore.query(alice.did, [{}]);
@@ -364,7 +363,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const sortedRecords = messages.sort((a,b) =>
@@ -385,7 +384,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           // get all of the records
@@ -403,7 +402,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const sortedRecords = messages.sort((a,b) =>
@@ -426,7 +425,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const sortedRecords = messages.sort((a,b) =>
@@ -450,7 +449,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const limit = 6;
@@ -479,7 +478,7 @@ export function testMessageStore(): void {
             messageTimestamp: TestDataGenerator.randomTimestamp()
           })));
           for (const message of messages) {
-            await messageStore.put(alice.did, message.message, await constructRecordsWriteIndexes(message.recordsWrite, true));
+            await messageStore.put(alice.did, message.message, await message.recordsWrite.constructRecordsWriteIndexes(true));
           }
 
           const limit = 4;
